# faster ALP encode