--- a/encodings/alp/src/lib.rs
+++ b/encodings/alp/src/lib.rs
@@ -1,3 +1,5 @@
+#![feature(convert_float_to_int)]
+
 pub use alp::*;
 pub use array::*;
 pub use compress::*;