--- a/encodings/alp/src/lib.rs
+++ b/encodings/alp/src/lib.rs
@@ -1,5 +1,3 @@
-#![feature(convert_float_to_int)]
-
 pub use alp::*;
 pub use array::*;
 pub use compress::*;